--- src/main/java/com/naekang/wealgo/domain/User.java
@@ -1,45 +1,0 @@
-package com.naekang.wealgo.domain;
-
-import com.naekang.wealgo.type.UserRole;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
-import org.springframework.security.crypto.password.PasswordEncoder;
-
-@Entity(name = "USERS")
-@Getter
-@Builder
-@ToString
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class User extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(unique = true)
-    private String email;
-
-    private String username;
-
-    private String password;
-
-    @Enumerated(EnumType.STRING)
-    private UserRole role;
-
-    public boolean checkPassword(PasswordEncoder passwordEncoder, String inputPassword) {
-        return passwordEncoder.matches(inputPassword, this.password);
-    }
-
-}
