+++ src/main/java/com/naekang/wealgo/domain/auth/dto/request/LoginRequestDTO.java
@@ -1,0 +1,23 @@
+package com.naekang.wealgo.domain.auth.dto.request;
+
+import javax.validation.constraints.Email;
+import javax.validation.constraints.NotBlank;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class LoginRequestDTO {
+
+    @Email(message = "이메일 형식이 맞지 않습니다.")
+    @NotBlank(message = "이메일을 입력해주세요.")
+    private String email;
+
+    @NotBlank(message = "비밀번호를 입력해주세요.")
+    private String password;
+
+}
