+++ src/main/java/com/naekang/wealgo/domain/auth/dto/request/SignUpRequestDTO.java
@@ -1,0 +1,35 @@
+package com.naekang.wealgo.domain.auth.dto.request;
+
+import com.naekang.wealgo.domain.auth.type.UserRole;
+import javax.validation.constraints.Email;
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.Pattern;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class SignUpRequestDTO {
+
+    @Email(message = "이메일 형식이 맞지 않습니다.")
+    @NotBlank(message = "이메일을 입력해주세요.")
+    private String email;
+
+    @NotBlank(message = "Solved.ac 닉네임을 입력해주세요.")
+    private String username;
+
+    @NotBlank(message = "비밀번호를 입력해주세요.")
+    @Pattern(regexp = "^(?=.*[A-Za-z])(?=.*\\d)(?=.*[@$!%*#?&])[A-Za-z\\d@$!%*#?&]{8,30}$",
+        message = "비밀번호는 8-30 자리이면서 1개 이상의 알파벳, 숫자, 특수문자를 포함해야합니다.")
+    private String password;
+
+    @NotBlank(message = "다시 한번 비밀번호를 입력해주세요.")
+    private String checkedPassword;
+
+    private UserRole role;
+
+}
