+++ src/main/java/com/naekang/wealgo/domain/auth/dto/response/SignUpResponseDTO.java
@@ -1,0 +1,12 @@
+package com.naekang.wealgo.domain.auth.dto.response;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class SignUpResponseDTO {
+
+    private String username;
+
+}
