+++ src/main/java/com/naekang/wealgo/domain/auth/repository/UserRepository.java
@@ -1,0 +1,11 @@
+package com.naekang.wealgo.domain.auth.repository;
+
+import com.naekang.wealgo.domain.auth.entity.User;
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    Optional<User> findByEmail(String email);
+
+}
