+++ src/main/java/com/naekang/wealgo/domain/auth/type/UserRole.java
@@ -1,0 +1,5 @@
+package com.naekang.wealgo.domain.auth.type;
+
+public enum UserRole {
+    USER, ADMIN;
+}
