+++ src/main/java/com/naekang/wealgo/domain/user/controller/UserController.java
@@ -1,0 +1,13 @@
+package com.naekang.wealgo.domain.user.controller;
+
+import com.naekang.wealgo.domain.user.service.UserService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+public class UserController {
+
+    private final UserService userService;
+
+}
