--- src/main/java/com/naekang/wealgo/dto/SignUpResponseDTO.java
@@ -1,12 +1,0 @@
-package com.naekang.wealgo.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class SignUpResponseDTO {
-
-    private String username;
-
-}
