--- src/main/java/com/naekang/wealgo/repository/UserRepository.java
@@ -1,11 +1,0 @@
-package com.naekang.wealgo.repository;
-
-import com.naekang.wealgo.domain.User;
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    Optional<User> findByEmail(String email);
-
-}
