--- src/main/java/com/naekang/wealgo/type/UserRole.java
@@ -1,5 +1,0 @@
-package com.naekang.wealgo.type;
-
-public enum UserRole {
-    USER, ADMIN;
-}
